Save data for one city on offline storage and shwo it there is no connection for internet

+++ app/src/main/java/com/example/shazahassan/forecastapp/weather/Weather.java
@@ -35,15 +35,15 @@ public class Weather {
     public class MainData {
         @SerializedName("temp")
         @Expose
-        public int temp;
+        public float temp;
 
         @SerializedName("pressure")
         @Expose
-        public int pressure;
+        public float pressure;
 
         @SerializedName("humidity")
         @Expose
-        public int humidity;
+        public float humidity;
     }
 
     public class SysData {
